fix: add use type switch in the navigation items
