+++ apps/vardast-client/src/app/(client)/profile/info/components/InfoPage.tsx
@@ -1,10 +1,7 @@
 "use client"
 
 import PageTitle from "@vardast/component/project/PageTitle"
-import { useRefreshUserMutation, UserType } from "@vardast/graphql/generated"
-import graphqlRequestClientWithToken from "@vardast/query/queryClients/graphqlRequestClientWithToken"
-import { ToggleGroup, ToggleGroupItem } from "@vardast/ui/toggle-group"
-import { clsx } from "clsx"
+import { UserType } from "@vardast/graphql/generated"
 import { useSession } from "next-auth/react"
 
 import UpdateInfoItem from "@/app/(client)/profile/info/components/UpdateInfoItem"
@@ -15,60 +12,10 @@ type InfoPageProps = {
 }
 
 const InfoPage = ({ isMobileView, title }: InfoPageProps) => {
-  const { data: session, update, status } = useSession()
-
-  const refreshUserMutation = useRefreshUserMutation(
-    graphqlRequestClientWithToken,
-    {
-      onSuccess: async (data) => {
-        await update({ ...session, ...data.refresh })
-      }
-    }
-  )
+  const { data: session } = useSession()
 
   return (
     <div className="flex h-full w-full flex-col gap md:gap-9">
-      <div className="pt md:ml-auto md:pt-0">
-        <ToggleGroup
-          className="flex gap-6"
-          type="single"
-          value={session?.type}
-          onValueChange={(value: UserType) => {
-            value &&
-              value !== session?.type &&
-              refreshUserMutation.mutate({
-                refreshInput: {
-                  accessToken: session?.accessToken,
-                  refreshToken: session?.refreshToken,
-                  type: value
-                }
-              })
-          }}
-          defaultValue={UserType.Real}
-        >
-          <ToggleGroupItem
-            disabled={status === "loading" || refreshUserMutation.isLoading}
-            className={clsx(
-              "py-1 text-alpha-500",
-              session?.type === UserType.Real && "!bg-primary !text-alpha-white"
-            )}
-            value={UserType.Real}
-          >
-            حقیقی
-          </ToggleGroupItem>
-          <ToggleGroupItem
-            disabled={status === "loading" || refreshUserMutation.isLoading}
-            className={clsx(
-              "py-1 text-alpha-500",
-              session?.type === UserType.Legal &&
-                "!bg-primary !text-alpha-white"
-            )}
-            value={UserType.Legal}
-          >
-            حقوقی
-          </ToggleGroupItem>
-        </ToggleGroup>
-      </div>
       {!isMobileView && <PageTitle title={title} />}
       <div className=" w-full gap-x  border-alpha-200 bg-alpha-white text-alpha-500 md:grid md:grid-cols-2">
         {session?.type === UserType.Real && (
